♻️ refactor: reorganize code into modules and remove unused imports/variables
